More system definition cleanups for ASDF 3.3
Most importantly, follow the naming conventions of secondary systems
so they don't cause ASDF to issue a warning.

Also, make asdf-system-connections not optional, since we probably
want those systems to always be defined. A yet cleaner alternative
would be to move those connections to a separate file
cl-graph-connections, but that's not backward compatible, so it's
not my decision to make.